Adjust plugin entry point to match -p tcms_pytest_plugin
