+++ setup.py
@@ -1,16 +1,20 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
+#
 # Copyright (c) 2019 Dmitry Dygalo <dadygalo@gmail.com>
+# Copyright (c) 2021 Bryan Mutai <mutaiwork@gmail.com>
+# Copyright (c) 2021 Alexander Todorov <atodorov@otb.bg>
+#
 # Licensed under the GPLv3: https://www.gnu.org/licenses/gpl.html
+
 from setuptools import setup
-import tcms_pytest_plugin
 
 README = open("README.rst").read()
 REQUIREMENTS = open("requirements.txt").readlines()
 
 setup(
     name="kiwitcms-pytest-plugin",
-    version=tcms_pytest_plugin.__version__,
+    version="0.1.0",
     author="Dmitry Dygalo",
     author_email="dadygalo@gmail.com",
     maintainer="Kiwi TCMS",
@@ -40,5 +44,5 @@ setup(
         "Topic :: Software Development :: Quality Assurance",
         "Topic :: Software Development :: Testing",
     ],
-    entry_points={"pytest11": ["tcms_pytest_plugin = tcms_pytest_plugin.plugin"]},
+    entry_points={"pytest11": ["tcms_pytest_plugin = tcms_pytest_plugin"]},
 )
